# Optimize PriorityQueue