``` ini

BenchmarkDotNet=v0.12.1, OS=Windows 10.0.19043
Intel Core i7-4790 CPU 3.60GHz (Haswell), 1 CPU, 8 logical and 4 physical cores
.NET Core SDK=6.0.100-preview.7.21379.14
  [Host]   : .NET Core 5.0.9 (CoreCLR 5.0.921.35908, CoreFX 5.0.921.35908), X64 RyuJIT
  ShortRun : .NET Core 3.1.18 (CoreCLR 4.700.21.35901, CoreFX 4.700.21.36305), X64 RyuJIT

Job=ShortRun  Toolchain=.NET Core 3.1  IterationCount=3  
LaunchCount=1  WarmupCount=3  

```
|      Method |       N |          Mean |          Error |        StdDev |
|------------ |-------- |--------------:|---------------:|--------------:|
|  **NewDefault** |    **1024** |      **65.21 μs** |       **2.229 μs** |      **0.122 μs** |
|  **NewDefault** |   **32768** |   **3,238.99 μs** |      **91.401 μs** |      **5.010 μs** |
|  **NewDefault** | **1048576** | **124,861.30 μs** |  **46,946.538 μs** |  **2,573.298 μs** |
|             |         |               |                |               |
|  **OldDefault** |    **1024** |      **93.65 μs** |       **7.599 μs** |      **0.417 μs** |
|  **OldDefault** |   **32768** |   **4,313.46 μs** |     **578.538 μs** |     **31.712 μs** |
|  **OldDefault** | **1048576** | **173,724.40 μs** |  **69,870.061 μs** |  **3,829.814 μs** |
|             |         |               |                |               |
| **NewComparer** |    **1024** |     **102.02 μs** |      **31.775 μs** |      **1.742 μs** |
| **NewComparer** |   **32768** |   **4,664.77 μs** |     **579.447 μs** |     **31.761 μs** |
| **NewComparer** | **1048576** | **236,077.04 μs** | **600,743.617 μs** | **32,928.785 μs** |
|             |         |               |                |               |
| **OldComparer** |    **1024** |      **91.81 μs** |       **5.915 μs** |      **0.324 μs** |
| **OldComparer** |   **32768** |   **4,558.61 μs** |     **935.439 μs** |     **51.275 μs** |
| **OldComparer** | **1048576** | **180,557.77 μs** | **117,233.991 μs** |  **6,425.991 μs** |